# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Expert Portal API
  version: 0.1.0
  description: REST contract for the Expert Portal MVP.
servers:
  - url: http://localhost:4000/api/v1
security:
  - bearerAuth: []
paths:
  /experts/me:
    get:
      summary: Get the authenticated expert profile
      responses:
        "200":
          description: Expert profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpertProfile"
    patch:
      summary: Update the authenticated expert profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpertProfileUpdateInput"
      responses:
        "200":
          description: Updated expert profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpertProfile"
  /experts/me/opportunities:
    get:
      summary: List opportunities for the authenticated expert
      responses:
        "200":
          description: Opportunity summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpportunitySummary"
  /opportunities/{opportunityId}:
    get:
      summary: Get opportunity detail
      parameters:
        - $ref: "#/components/parameters/OpportunityId"
      responses:
        "200":
          description: Opportunity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityDetail"
  /opportunities/{opportunityId}/accept:
    post:
      summary: Accept an opportunity
      parameters:
        - $ref: "#/components/parameters/OpportunityId"
      responses:
        "200":
          description: Updated opportunity summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunitySummary"
  /opportunities/{opportunityId}/decline:
    post:
      summary: Decline an opportunity
      parameters:
        - $ref: "#/components/parameters/OpportunityId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Updated opportunity summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunitySummary"
  /opportunities/{opportunityId}/screening-answers:
    post:
      summary: Submit screening answers
      parameters:
        - $ref: "#/components/parameters/OpportunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  items:
                    $ref: "#/components/schemas/ScreeningAnswerInput"
                submit:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Submitted answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreeningAnswer"
  /availability-rules:
    get:
      summary: List availability rules
      responses:
        "200":
          description: Availability rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilityRule"
    post:
      summary: Create availability rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityRuleInput"
      responses:
        "201":
          description: Created availability rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityRule"
    patch:
      summary: Update availability rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/AvailabilityRuleInput"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: string
                      format: uuid
      responses:
        "200":
          description: Updated availability rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityRule"
  /bookings:
    get:
      summary: List bookings for the authenticated expert
      responses:
        "200":
          description: Booking list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/ics:
    post:
      summary: Export a booking as an ICS calendar entry
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Generated ICS content
          content:
            text/calendar:
              schema:
                type: string
  /compliance/attestations:
    get:
      summary: List compliance attestations
      responses:
        "200":
          description: Attestations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceAttestation"
    post:
      summary: Complete a compliance attestation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ruleId, accepted]
              properties:
                ruleId:
                  type: string
                  format: uuid
                accepted:
                  type: boolean
      responses:
        "201":
          description: Created attestation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceAttestation"
  /payments:
    get:
      summary: List payment records
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRecord"
  /payments/{engagementId}/external-link:
    post:
      summary: Generate a secure external payment link
      parameters:
        - name: engagementId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: External payment link
          content:
            application/json:
              schema:
                type: object
                required: [url, expiresAt]
                properties:
                  url:
                    type: string
                    format: uri
                  expiresAt:
                    type: string
                    format: date-time
  /ops/opportunities:
    post:
      summary: Create an opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "201":
          description: Created opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityDetail"
    patch:
      summary: Update an opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/OpportunityInput"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: string
                      format: uuid
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityDetail"
  /ops/opportunities/{opportunityId}/candidates:
    post:
      summary: Add expert candidates to an opportunity
      parameters:
        - $ref: "#/components/parameters/OpportunityId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expertIds]
              properties:
                expertIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Updated candidate list
          content:
            application/json:
              schema:
                type: object
                required: [opportunityId, expertIds]
                properties:
                  opportunityId:
                    type: string
                    format: uuid
                  expertIds:
                    type: array
                    items:
                      type: string
                      format: uuid
  /ops/verification/{expertId}:
    patch:
      summary: Update verification state for an expert
      parameters:
        - $ref: "#/components/parameters/ExpertId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/VerificationStatus"
                reason:
                  type: string
      responses:
        "200":
          description: Updated verification summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationSummary"
  /ops/compliance-rules:
    post:
      summary: Create a compliance rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplianceRuleInput"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceRule"
    patch:
      summary: Update a compliance rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ComplianceRuleInput"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: string
                      format: uuid
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceRule"
  /ops/payments/{paymentId}/status:
    patch:
      summary: Update payment status
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/PaymentStatus"
                note:
                  type: string
      responses:
        "200":
          description: Updated payment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRecord"
  /ops/audit-events:
    get:
      summary: List audit events
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEvent"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OpportunityId:
      name: opportunityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BookingId:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ExpertId:
      name: expertId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ExpertStatus:
      type: string
      enum: [active, pending_verification, restricted, suspended]
    VerificationStatus:
      type: string
      enum: [not_started, pending, completed, failed, expired]
    BookingStatus:
      type: string
      enum: [proposed, scheduled, completed, cancelled, no_show]
    PaymentStatus:
      type: string
      enum: [not_eligible, pending_request, requested_external, submitted, approved, paid, held]
    OpportunityStatus:
      type: string
      enum: [invited, screening_in_progress, accepted, declined, slot_proposed, scheduled, completed, payment_pending, paid, held]
    VerificationSummary:
      type: object
      required: [status, personaInquiryId, updatedAt]
      properties:
        status:
          $ref: "#/components/schemas/VerificationStatus"
        personaInquiryId:
          type: string
        updatedAt:
          type: string
          format: date-time
        outcomeSummary:
          type: string
    ExpertProfile:
      type: object
      required:
        [id, email, status, verification, fullName, headline, timezone, expertiseTags, roleHistory, recencySignals, lastConfirmedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        status:
          $ref: "#/components/schemas/ExpertStatus"
        verification:
          $ref: "#/components/schemas/VerificationSummary"
        fullName:
          type: string
        headline:
          type: string
        timezone:
          type: string
        expertiseTags:
          type: array
          items:
            type: string
        roleHistory:
          type: array
          items:
            $ref: "#/components/schemas/RoleHistoryItem"
        recencySignals:
          type: array
          items:
            $ref: "#/components/schemas/RecencySignal"
        lastConfirmedAt:
          type: string
          format: date-time
    ExpertProfileUpdateInput:
      type: object
      properties:
        fullName:
          type: string
        headline:
          type: string
        timezone:
          type: string
        expertiseTags:
          type: array
          items:
            type: string
    RoleHistoryItem:
      type: object
      required: [company, title, startedAt]
      properties:
        company:
          type: string
        title:
          type: string
        startedAt:
          type: string
          format: date
        endedAt:
          type: string
          format: date
    RecencySignal:
      type: object
      required: [id, topic, lastHandsOnDate, confidence, source]
      properties:
        id:
          type: string
          format: uuid
        topic:
          type: string
        lastHandsOnDate:
          type: string
          format: date
        confidence:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
    ScreeningQuestion:
      type: object
      required: [id, prompt, required]
      properties:
        id:
          type: string
          format: uuid
        prompt:
          type: string
        required:
          type: boolean
    ScreeningAnswerInput:
      type: object
      required: [questionId, answerText]
      properties:
        questionId:
          type: string
          format: uuid
        answerText:
          type: string
    ScreeningAnswer:
      type: object
      required: [id, questionId, answerText, submittedAt, version]
      properties:
        id:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        answerText:
          type: string
        submittedAt:
          type: string
          format: date-time
        version:
          type: integer
          minimum: 1
    OpportunitySummary:
      type: object
      required: [id, title, topic, status, deadlineAt, rateUsd, complianceFlags]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        topic:
          type: string
        status:
          $ref: "#/components/schemas/OpportunityStatus"
        deadlineAt:
          type: string
          format: date-time
        rateUsd:
          type: integer
        complianceFlags:
          type: array
          items:
            type: string
    OpportunityDetail:
      allOf:
        - $ref: "#/components/schemas/OpportunitySummary"
        - type: object
          required: [description, estimatedMinutes, screeningQuestions]
          properties:
            description:
              type: string
            estimatedMinutes:
              type: integer
            screeningQuestions:
              type: array
              items:
                $ref: "#/components/schemas/ScreeningQuestion"
    OpportunityInput:
      type: object
      required: [title, topic, deadlineAt, description, estimatedMinutes, rateUsd]
      properties:
        title:
          type: string
        topic:
          type: string
        deadlineAt:
          type: string
          format: date-time
        description:
          type: string
        estimatedMinutes:
          type: integer
        rateUsd:
          type: integer
    AvailabilityRule:
      type: object
      required: [id, timezone, weekday, startTime, endTime, leadHours, blackoutDates]
      properties:
        id:
          type: string
          format: uuid
        timezone:
          type: string
        weekday:
          type: integer
          minimum: 0
          maximum: 6
        startTime:
          type: string
        endTime:
          type: string
        leadHours:
          type: integer
        blackoutDates:
          type: array
          items:
            type: string
            format: date
    AvailabilityRuleInput:
      type: object
      required: [timezone, weekday, startTime, endTime, leadHours]
      properties:
        timezone:
          type: string
        weekday:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        leadHours:
          type: integer
        blackoutDates:
          type: array
          items:
            type: string
            format: date
    Booking:
      type: object
      required: [id, opportunityId, expertId, title, startAt, endAt, status, timezone]
      properties:
        id:
          type: string
          format: uuid
        opportunityId:
          type: string
          format: uuid
        expertId:
          type: string
          format: uuid
        title:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/BookingStatus"
        timezone:
          type: string
    ComplianceRule:
      type: object
      required: [id, name, ruleType, description, blocking]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ruleType:
          type: string
        description:
          type: string
        blocking:
          type: boolean
    ComplianceRuleInput:
      type: object
      required: [name, ruleType, description, blocking]
      properties:
        name:
          type: string
        ruleType:
          type: string
        description:
          type: string
        blocking:
          type: boolean
    ComplianceAttestation:
      type: object
      required: [id, ruleId, accepted, createdAt]
      properties:
        id:
          type: string
          format: uuid
        ruleId:
          type: string
          format: uuid
        accepted:
          type: boolean
        createdAt:
          type: string
          format: date-time
    PaymentRecord:
      type: object
      required: [id, engagementId, status, amountUsd, dueAt, payoutLinkIssued]
      properties:
        id:
          type: string
          format: uuid
        engagementId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/PaymentStatus"
        amountUsd:
          type: integer
        dueAt:
          type: string
          format: date-time
        payoutLinkIssued:
          type: boolean
    AuditEvent:
      type: object
      required: [id, actorType, actorId, action, entityType, entityId, payload, createdAt]
      properties:
        id:
          type: string
          format: uuid
        actorType:
          type: string
        actorId:
          type: string
          format: uuid
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time